test(hashing): refresh edge case tests for hash maps and sets

+++ tests/cpp/test_hashing.cpp
@@ -1,10 +1,11 @@
 #include "test_common.hpp"
 #include "../../cpp/hashing/hashing.cpp"
 
-int main() {
+int main()
+{
     std::vector<std::string> strs{"eat", "tea", "tan", "ate", "nat", "bat"};
     auto groups = dsa::groupAnagrams(strs);
-    CHECK_EQ(groups.size(), 3);
-    CHECK_EQ(dsa::longestConsecutive({100, 4, 200, 1, 3, 2}), 4);
+    CHECK_EQ((static_cast<long long>(groups.size())), (3));
+    CHECK_EQ((dsa::longestConsecutive({100, 4, 200, 1, 3, 2})), (4));
     return 0;
 }
